requirements.txt

streamlit==1.28.0
pandas==2.0.3
scikit-learn==1.3.0
requests
python-dotenv